refactor: shrink ChatListView code
